kernel review: hot-path performance; drop libgcc so rv64 runs on -cpu rv64
- memcpy/memmove/memset move 8 bytes at a time when source and
  destination share alignment (byte head/tail otherwise), keeping every
  access naturally aligned for arm64's -mstrict-align. They were
  byte-at-a-time and sit under every struct copy/zeroing and the EtherCAT
  frame copies. no-tree-loop-distribute-patterns stops GCC turning the
  loops back into self-calls. New `test mem` subtest checks all three
  (incl. overlapping memmove both ways and guard bytes) against a byte
  reference over every alignment; it's part of `test all`.
- Scheduler: the equal-deadline tie-break used a timestamp, costing a timer
  read + 128-bit divide on every context switch; only the order matters, so
  it's now a relaxed global sequence counter.
- util::mul_div_u64 computes (a/c)*b + ((a%c)*b)/c, which is exact without
  a 128-bit divide for every tick<->time conversion (inline long-division
  fallback for the rare overflow case). cnc/interpreter's path-length sums
  accumulate in double (they end in a float sqrt anyway) instead of
  unsigned __int128.

With those, neither arch links any libgcc object, so rv64 no longer needs
Ubuntu 26.04's RVA23-built libgcc or `-cpu rva23s64`: it boots and passes
`test all` on the Makefile's plain `-cpu rv64`. CLAUDE.md updated.

Both arches: banner, ci-ok, [ec0] cycle=250us, `test all` 13/0 (3/3 each;
rv64 on -cpu rv64).

+++ cnc/interpreter.cpp
@@ -471,18 +471,20 @@ static uint32_t feed_cps_for_state(const Runtime::ChannelState& state,
 static uint32_t combined_path_counts(const int32_t start[motion::MAX_AXES],
                                      const int32_t targets[motion::MAX_AXES],
                                      uint64_t axis_mask) noexcept {
-    // Accumulate the squared deltas in 128-bit. A single full-range axis
-    // delta (~4.3e9 counts) squared is ~1.8e19, which overflows the int64
-    // multiply (signed UB) the old code used; summing several axes overflows
-    // uint64 too. unsigned __int128 holds it exactly for any axis count.
-    unsigned __int128 sum_sq = 0;
+    // Accumulate the squared deltas in double. A single full-range axis
+    // delta (~4.3e9 counts) squared is ~1.8e19, which overflows int64 (the
+    // original code's signed-UB bug) and, summed over axes, uint64. double
+    // holds each square to ~1e-16 relative error — far finer than the float
+    // sqrt the result goes through — without the 128-bit arithmetic that
+    // pulled libgcc's __floatuntidf in.
+    double sum_sq = 0.0;
     for (size_t i = 0; i < motion::MAX_AXES; ++i) {
         if ((axis_mask & (1ull << i)) == 0) continue;
         const int64_t d = static_cast<int64_t>(targets[i]) - start[i];
-        const uint64_t ad = d < 0 ? static_cast<uint64_t>(-d) : static_cast<uint64_t>(d);
-        sum_sq += static_cast<unsigned __int128>(ad) * ad;
+        const double ad = static_cast<double>(d < 0 ? -d : d);
+        sum_sq += ad * ad;
     }
-    return static_cast<uint32_t>(sqrt_approx(static_cast<float>(static_cast<double>(sum_sq))));
+    return static_cast<uint32_t>(sqrt_approx(static_cast<float>(sum_sq)));
 }
 
 static int axis_offset_slot(char word) {
@@ -875,14 +877,14 @@ static bool execute_axes(Runtime::ChannelState& state,
                 const int64_t dx = static_cast<int64_t>(tip_targets[0]) - state.tool_tip_pos[0];
                 const int64_t dy = static_cast<int64_t>(tip_targets[1]) - state.tool_tip_pos[1];
                 const int64_t dz = static_cast<int64_t>(tip_targets[2]) - state.tool_tip_pos[2];
-                // 128-bit accumulate — see combined_path_counts: each squared
+                // double accumulate — see combined_path_counts: each squared
                 // delta can exceed int64/uint64 range at full travel.
-                auto sq = [](int64_t v) -> unsigned __int128 {
-                    const uint64_t a = v < 0 ? static_cast<uint64_t>(-v) : static_cast<uint64_t>(v);
-                    return static_cast<unsigned __int128>(a) * a;
+                auto sq = [](int64_t v) -> double {
+                    const double a = static_cast<double>(v < 0 ? -v : v);
+                    return a * a;
                 };
-                const unsigned __int128 sum_sq = sq(dx) + sq(dy) + sq(dz);
-                path_counts = static_cast<uint32_t>(sqrt_approx(static_cast<float>(static_cast<double>(sum_sq))));
+                const double sum_sq = sq(dx) + sq(dy) + sq(dz);
+                path_counts = static_cast<uint32_t>(sqrt_approx(static_cast<float>(sum_sq)));
             } else {
                 path_counts = combined_path_counts(state.targets, targets, axis_mask);
             }
